Refactor `PlayerStatus` to support remote players
`PlayerStatus` now has a vector for local players and a map of vectors for remote users and their players. Remote users are saved under "tux_remote" entries, where "nickname" and "id" entries are present.

Additionally:

* Simplifies the `Player` constructor.
* `LevelIntro` now holds a vector of pointers to added `Player`s.
* `network::UserProtocol` now holds a string variable which stores the nickname the host goes under.

+++ src/supertux/game_network_server_user.hpp
@@ -35,7 +35,7 @@ public:
 
   void write(Writer& writer) const override;
 
-  size_t get_num_players() const { return player_controllers.size(); }
+  int get_num_players() const { return static_cast<int>(player_controllers.size()); }
 
 public:
   // Used by servers to manage remote player input.
